asset_management/requirements.txt
Django==5.2.13
Pillow==12.2.0
django-import-export==4.4.0
openpyxl==3.1.5
xlrd==2.0.2
django-crispy-forms==2.6
crispy-bootstrap5==2026.3
tablib==3.9.0
